requirements.txt
Telethon==1.34.0
discord-webhook==1.3.1
PyYAML